Conflicts:
	script_src/python/trackpy/cpp_wrapper.py

kept new way of dealing with the DB if there is nothing in it

Still has demo paths for binary locations

+++ iden/cxx/data_proc1.cpp
@@ -142,7 +142,7 @@ void iden::RecenterImage(Image2D &image)
 // tac 2009-09-15
 // 
 IppStatus iden::FindLocalMax_2D(Image2D &image_bpass, Image2D &image_bpass_thresh, Image2D &image_subtracted,
-				const int intensity_threshold, const int dilation_radius)
+				const float intensity_threshold, const int dilation_radius)
 {
   IppStatus status;
   Image2D image_dilated(image_bpass.get_height(), image_bpass.get_width());
